Added kill-ihaskell helper
This kills the buffers associated with the Jupyter process, prompting
the user when the processes are alive (or at least undead!).

This is helpful because sometimes IHaskell crashes, apparently due to
holding too many file handles, leaving the various Jupyter processes in
an undead state where they need to be manually killed.