Don't evaluate module definitions in IHaskell
This is a gross hack. IHaskell has rudimentary support for defining an
entire module in a single block so that it can be compiled to object
code. This interacts poorly with wanting a notebook file to also work as
a regular Haskell module definition.

The problem is that everything in the source block with the "module"
statement is in a separate module from the rest of the notebook. If we
leave it in its own source block, it doesn't do anything useful, but it
is saved to disk, which gets outorg a bit confused about what file it is
working with.

To solve this, the my/org-execute-up-to-point helper ignores a block
that includes a module definition.