A selection of packages specifically built with clang
These packages expose C++ stdlib types in their APIs, so they must be
built with libc++ if a dependent program is to be built with libc++
and linked with them.

Ideally, such a program would operate entirely within a nixpkgs tree
built against libc++, but that is not easily done (e.g. many packages use
stdenv.isDarwin as a proxy for building with clang).