Trying out direnv
The emacs integration would be terrific, but it seems there are some
rough spots with precisely when direnv updates emacs's paths when
switching into a new buffer. This can interact poorly with compiler
servers such as LSP tools.

On the other hand, it makes finding a compiler like `clang++` from a
`shell.nix` work when evaluating a `src` block in an org file.