Disable the maximum size on isync messages
This setting causes isync / mbsync to skip messages with attachments
above a certain size. The wanted behavior is to just not download the
attachment, but perhaps indicate that there is a missing
attachment. Skipping the message altogether means that you do not
receive the email at all. I am disabling it for now and relying on
relatively conservative maximum attachment size settings on various
email servers.