Experimented with helm in its own frame
It looked nice sometimes, but ultimately I found the lack of
distinction between the helm popup and the original frame a bit
visually noisy. The location of the popup is, by default, at point,
which I also found a bit irritating as I'm not always looking at point
when invoking a command. In such a case, having the helm interface
appear at an easy-to-locate position is beneficial.