Use doom-modeline instead of moody
It requires less customization from me, and has some nice icons.

I include a patch to always show an indicator for the god-mode state,
and I made that indicator a filled circle rather than the default
"<G>". I also patched the "project" modeline so that the indicator
appears in magit and dashboard buffers.