Patch lsp-mode and company-lsp for unusual completions
When company tries to complete a Haskell qualified import (e.g. `H.`,
where we have `import import qualified Numeric.LinearAlgebra as H` in
scope), we get an error because some datum is assumed to be a hash
table, but it is not. These patches check that an argument is a hash
table before applying hash table functions to it. If it is not a hash
table, we make a best effort to provide what is needed.