Custom eval-last-sexp that inserts an arrow before the result
When called with a prefix, eval-last-sexp inserts the results of
evaluating the sexp before point into the buffer. This new function
does the same, but inserts an arrow character between the sexp and the
results of evaluation. This improves readability when you want to keep
both the sexp and its result in some text.