emacs: org helper to re-evaluate all blocks in a session
Executes all org src blocks up to the current point with a :session header whose value is
something other than "none".

This does not look at the specific value of the :session header or the
language or the :results header, it just executes everything that is
part of some session.