emacs: an attempt to get direnv updating before lsp tries to start
Sometimes the emacs mode tries to start the LSP server before
direnv has had a chance to update environment variables. One solution
is to open a file in a project directory that does not need an LSP
server (I have an Org file of notes in every project that I use for
this purpose) as this gives direnv a chance to run.

This tweak inserts the direnv update earlier in the file opening process.