emacs: helper for adding a book with org-books
I put all my read books under one heading (perhaps not a great idea!),
so having to enter it each time I add a book is a nuisance. This
helper finds the `Books` heading in the `org-books-file` and adds the
book there.

A minor note is that org-books hard-codes the use of helm in the
heading selection which can be a bit jarring since the author
selection uses standard completing-read.