Tweak to insert \( \) pairs in org-mode when hitting a $
The $ character used to delimit inline latex in org-mode can cause
issues with some exporters (I think HTML is the main culprit), and I
believe parsing may not be entirely reliable. What works somewhat
better is surrounding inline latex with \( \), but these are arguably
a bit more effort to type, or at least don't take advantage of the
muscle memory for typing $. This tweak makes the $ key insert a \( \)
pair and put point in between them. To insert an actual $ character,
hit another $ after the first: The \( \) pair will be deleted, and a
literal $ inserted.