emacs: render time stamps with a fixed-width face
I sometimes put time stamps (or date stamps) in headlines. When these
headlines appear in a clock table, the table formatting is thrown off
because while tables are rendered with fixed-width faces, the variable
width face for dates overrides that. This change fixes the issue in
tables at the expense of dates everywhere being rendered with a fixed
width face.