emacs: disable mode line face height config
This is not an issue at the moment given changes in emacs git. The
problem is when using a proportional font for the mode line, the
height can be more than I want in the mode line. The monospace font
size is tighter, as that is what I use elsewhere.