emacs: bind `C-,` to `flyspell-correct-previous`
Something that happens to me quite often is that I will be typing
along, and notice that flyspell has highlighted a misspelling a few
words back. What I want is to be able to jump straight to that error
and bring up the spelling correction UI. This convenient keybinding
now does just that.