emacs: change org-modern-star symbols
A problem I have with the recommended set is that they do not indent
very well (at least with the fonts I use). This leads to stars being
aligned vertically, when I do want some indentation to make it easier
to attach more-indented items to their parent as distinct from a
less-indented item just below.