emacs: use plist serialization with lsp-mode
GIven that we pre-compile lsp-mode when building emacs, I'm not sure
if this environment variable is necessary. It is recommended in the
instructions. This setting is aimed to make lsp-mode somewhat faster,
and it is also necessary to make use of the emacs-lsp-booster wrapper
program that aims to make LSP interactions from emacs more efficient.